+++ pytorch_lightning/trainer/configuration_validator.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 import pytorch_lightning as pl
 from pytorch_lightning.trainer.states import TrainerFn
-from pytorch_lightning.utilities import rank_zero_warn
+from pytorch_lightning.utilities import rank_zero_warn, rank_zero_deprecation
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
 from pytorch_lightning.utilities.model_helpers import is_overridden
 
@@ -86,6 +86,21 @@ class ConfigValidator:
                 "When using `Trainer(accumulate_grad_batches != 1)` and overriding"
                 "`LightningModule.optimizer_{step,zero_grad}`, the hooks will not be called on every batch"
                 "(rather, they are called on every optimization step)."
+            )
+
+    def check_progress_bar(self, model: "pl.LightningModule") -> None:
+        r"""
+        Checks if get_progress_bar_dict is overriden and sends a deprecation warning.
+
+        Args:
+            model: The model to check the get_progress_bar_dict method.
+
+        """
+        if is_overridden("get_progress_bar_dict", model):
+            rank_zero_deprecation(
+                "The `LightningModule.get_progress_bar_dict` method was deprecated in v1.5 and will be removed in v1.7."
+                "Please use the `trainer.progress_bar_callback.get_progress_bar_dict` instead.",
+                stacklevel=5,
             )
 
     def __verify_eval_loop_configuration(self, model: "pl.LightningModule", stage: str) -> None: