+++ pytorch_lightning/trainer/properties.py
@@ -39,7 +39,7 @@ from pytorch_lightning.trainer.connectors.checkpoint_connector import Checkpoint
 from pytorch_lightning.trainer.connectors.logger_connector import LoggerConnector
 from pytorch_lightning.trainer.connectors.logger_connector.result import ResultCollection
 from pytorch_lightning.trainer.states import RunningStage, TrainerFn, TrainerState, TrainerStatus
-from pytorch_lightning.utilities import DeviceType, DistributedType, rank_zero_deprecation, rank_zero_warn
+from pytorch_lightning.utilities import DeviceType, DistributedType, rank_zero_deprecation
 from pytorch_lightning.utilities.argparse import (
     add_argparse_args,
     from_argparse_args,
@@ -298,19 +298,11 @@ class TrainerProperties(ABC):
         """Read-only for progress bar metrics."""
         ref_model = self.lightning_module
         ref_model = cast(pl.LightningModule, ref_model)
-
-        standard_metrics = self.progress_bar_callback.get_progress_bar_dict(ref_model)
-        pbar_metrics = self.progress_bar_metrics
-        duplicates = list(standard_metrics.keys() & pbar_metrics.keys())
-        if duplicates:
-            rank_zero_warn(
-                f"The progress bar already tracks a metric with the name(s) '{', '.join(duplicates)}' and"
-                f" `self.log('{duplicates[0]}', ..., prog_bar=True)` will overwrite this value. "
-                " If this is undesired, change the name or override `get_progress_bar_dict()`"
-                " in `LightingModule`.",
-                UserWarning,
-            )
-        return {**standard_metrics, **pbar_metrics}
+        rank_zero_deprecation(
+            "`trainer.progress_bar_dict` is deprecated in v1.5 and will be removed in v1.7."
+            " Use `trainer.progress_bar_callback.get_progress_bar_dict` instead."
+        )
+        return self.progress_bar_callback.get_progress_bar_dict(self, ref_model)
 
     @property
     def _should_reload_dl_epoch(self) -> bool: