+++ tests/callbacks/test_progress_bar.py
@@ -562,8 +562,8 @@ def _test_progress_bar_max_val_check_interval(
 
 def test_get_progress_bar_dict_callback(tmpdir):
     class TestProgressBar(ProgressBar):
-        def get_progress_bar_dict(self, model):
-            items = super().get_progress_bar_dict(model)
+        def get_progress_bar_dict(self, trainer, model):
+            items = super().get_progress_bar_dict(trainer, model)
             items.pop("v_num", None)
             return items
 
@@ -576,7 +576,7 @@ def test_get_progress_bar_dict_callback(tmpdir):
     )
     model = BoringModel()
     model.truncated_bptt_steps = 2
-    standard_metrics = trainer.progress_bar_callback.get_progress_bar_dict(model)
+    standard_metrics = trainer.progress_bar_callback.get_progress_bar_dict(trainer, model)
     assert "loss" in standard_metrics.keys()
     assert "split_idx" in standard_metrics.keys()
     assert "v_num" not in standard_metrics.keys()